+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## 0.1.0
 
-- Migrate to null safety
+- Migrate to null safety.
+- Migrate `example` app to null safety.
 - Improve usability of API using relative values for `depth` and `velocity`.
-- Replace `LitStarfieldBackground` with `LitStarfieldContainer`
+- Replace `LitStarfieldBackground` with `LitStarfieldContainer`.
+- Add `model.dart`, `view.dart` and `controller.dart` sub libraries.
+- Add customization options on `LitStarfieldContainer`.
 
 ## 0.0.1