+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1
+
+- Adjust `description` on `pubspec.yaml`.
+
 ## 0.1.0
 
 - Migrate to null safety.